ebackend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.1.1
django-clearcache==1.2.1
django-cors-headers==3.7.0
django-environ==0.8.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
encodedecode==0.0.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.1
style==1.1.0
tzdata==2022.2
update==0.0.1
urllib3==1.26.8
whitenoise==5.3.0
